Move overload to stub file
